+++ package/prefixdb/prefixdb/multidb/options/backup_config_json.hpp
@@ -12,7 +12,6 @@ struct backup_config_json
   JSON_NAME(depth)
   JSON_NAME(start_time)
   JSON_NAME(period_s)
-  /*JSON_NAME(start_delay_s)*/
 
   typedef wfc::json::object<
     backup_config,
@@ -20,7 +19,7 @@ struct backup_config_json
       wfc::json::member<n_enabled,       backup_config, bool,        &backup_config::enabled>,
       wfc::json::member<n_period_s,      backup_config, time_t,      &backup_config::period_s>,
       wfc::json::member<n_start_time,    backup_config, std::string, &backup_config::start_time>,
-      wfc::json::member<n_depth,         backup_config, size_t,      &backup_config::depth>,
+      wfc::json::member<n_depth,         backup_config, uint32_t,      &backup_config::depth>,
       wfc::json::member<n_path,          backup_config, std::string, &backup_config::path>
     >,
     wfc::json::strict_mode
